recommendations/requirements.txt
grpc-interceptor ~=0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4